+++ NEWS.md
@@ -2,6 +2,7 @@
 
 Initial port of `httptest` to work on top of `httr2`. All APIs preserved except:
 
+* Mock file hashes and error messages for multipart or file upload requests have changed.
 * `save_response()` requires a `file` path argument because `httr2_response` objects do not contain their `request`, which is needed to construct the mock file path
 * Redacting functions are applied to both the `httr2_response` and `httr2_request` objects, the latter in order to be able to alter the URL/mock file path being written. As a result, `gsub_request()` and `set_requester()` were removed because they are no longer necessary.
 * `redact_auth()`, which had previously been reduced to an alias for `redact_cookies()`, was removed